update 6.9.32 - commands in one cog file
