+++ utils/startup.py
@@ -15,18 +15,20 @@ def get_config():
 
 def load_commands(bot):
 	blacklisted_cogs = []
-	cogs = [
-		'help',
-		'economy'
-	]
+	commands_dir = "cogs/commands/"
 
-	for cog in cogs:
+	for cog in os.listdir(commands_dir):
+		if cog.endswith(".py"):
+			cog = cog[:-3]
+		else:
+			continue
+
 		if cog in blacklisted_cogs:
 			print(f"{t_yellow}Skipped cog: {cog} {t_white}")
 		else:
 			try:
-				bot.load_extension(f"cogs.commands.{cog}")
-				print(f"Loaded cog: {cog}")
+				bot.load_extension(f"{commands_dir.replace('/', '.')}{cog}")
+				print(f"Loaded command cog: {cog}")
 			except Exception as e:
 				exception = f"{type(e).__name__}: {e}"
 				print(f"{t_red}Failed to load cog: cogs/commands/{cog}\n{exception} {t_white}")
@@ -34,20 +36,19 @@ def load_commands(bot):
 
 def load_events(bot):
 	blacklisted_cogs = []
-	cogs = [
-		'on_command_completion',
-		'on_command_error',
-		'on_guild_join',
-		'on_guild_remove',
-		'on_ready'
-	]
+	events_dir = "cogs/events/"
 
-	for cog in cogs:
+	for cog in os.listdir(events_dir):
+		if cog.endswith(".py"):
+			cog = cog[:-3]
+		else:
+			continue
+
 		if cog in blacklisted_cogs:
-			print(f"{t_yellow}Skipped cog: {cog} {t_white}")
+			print(f"{t_yellow}Skipped event: {cog} {t_white}")
 		else:
 			try:
-				bot.load_extension(f"cogs.commands.{cog}")
+				bot.load_extension(f"{events_dir.replace('/', '.')}{cog}")
 				print(f"Loaded cog: {cog}")
 			except Exception as e:
 				exception = f"{type(e).__name__}: {e}"
